slice pushdown optimization

+++ polars/polars-lazy/src/physical_plan/executors/slice.rs
@@ -5,12 +5,12 @@ use polars_core::prelude::*;
 pub struct SliceExec {
     pub input: Box<dyn Executor>,
     pub offset: i64,
-    pub len: usize,
+    pub len: u32,
 }
 
 impl Executor for SliceExec {
     fn execute(&mut self, state: &ExecutionState) -> Result<DataFrame> {
         let df = self.input.execute(state)?;
-        Ok(df.slice(self.offset, self.len))
+        Ok(df.slice(self.offset, self.len as usize))
     }
 }
